+++ src/frontend/unions.rs
@@ -27,6 +27,7 @@ pub(super) fn parse(parser: &mut Parser, source: &mut Source) -> Result<(), Faul
             variants.push(Ty::Record(case.text.clone()));
             source.records.push(Record {
                 generic_parameters: Vec::new(),
+                generic_constraints: Vec::new(),
                 name: case,
                 fields,
                 is_class: false,